Introduce InventoryLowQuantityAdminUi module. Initial
