+++ app/code/Magento/InventoryLowQuantityNotification/etc/module.xml
@@ -6,11 +6,5 @@
  */
 -->
 <config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:Module/etc/module.xsd">
-    <module name="Magento_InventoryLowQuantityNotification" setup_version="1.0.0" >
-        <sequence>
-            <module name="Magento_Catalog"/>
-            <module name="Magento_Inventory"/>
-            <module name="Magento_InventoryCatalog"/>
-        </sequence>
-    </module>
+    <module name="Magento_InventoryLowQuantityNotification" setup_version="1.0.0" />
 </config>
