+++ app/code/Magento/InventoryLowQuantityNotificationAdminUi/Observer/ProcessSourceItemConfigurationsObserver.php
@@ -1,0 +1,69 @@
+<?php
+/**
+ * Copyright © Magento, Inc. All rights reserved.
+ * See COPYING.txt for license details.
+ */
+declare(strict_types=1);
+
+namespace Magento\InventoryLowQuantityNotificationAdminUi\Observer;
+
+use Magento\Catalog\Api\Data\ProductInterface;
+use Magento\Catalog\Controller\Adminhtml\Product\Save;
+use Magento\Framework\Event\ObserverInterface;
+use Magento\Framework\Event\Observer as EventObserver;
+use Magento\InventoryConfigurationApi\Model\IsSourceItemManagementAllowedForProductTypeInterface;
+use Magento\InventoryLowQuantityNotificationAdminUi\Model\SourceItemsConfigurationProcessor;
+
+/**
+ * Save source relations (configuration) during product persistence via controller
+ *
+ * This needs to be handled in dedicated observer, because there is no pre-defined way of making several API calls for
+ * Form submission handling
+ */
+class ProcessSourceItemConfigurationsObserver implements ObserverInterface
+{
+    /**
+     * @var IsSourceItemManagementAllowedForProductTypeInterface
+     */
+    private $isSourceItemManagementAllowedForProductType;
+
+    /**
+     * @var SourceItemsConfigurationProcessor
+     */
+    private $sourceItemsConfigurationProcessor;
+
+    /**
+     * @param IsSourceItemManagementAllowedForProductTypeInterface $isSourceItemManagementAllowedForProductType
+     * @param SourceItemsConfigurationProcessor $sourceItemsConfigurationProcessor
+     */
+    public function __construct(
+        IsSourceItemManagementAllowedForProductTypeInterface $isSourceItemManagementAllowedForProductType,
+        SourceItemsConfigurationProcessor $sourceItemsConfigurationProcessor
+    ) {
+        $this->isSourceItemManagementAllowedForProductType = $isSourceItemManagementAllowedForProductType;
+        $this->sourceItemsConfigurationProcessor = $sourceItemsConfigurationProcessor;
+    }
+
+    /**
+     * @param EventObserver $observer
+     * @return void
+     */
+    public function execute(EventObserver $observer)
+    {
+        /** @var ProductInterface $product */
+        $product = $observer->getEvent()->getProduct();
+        if ($this->isSourceItemManagementAllowedForProductType->execute($product->getTypeId()) === false) {
+            return;
+        }
+
+        /** @var Save $controller */
+        $controller = $observer->getEvent()->getController();
+
+        $sources = $controller->getRequest()->getParam('sources', []);
+        $assignedSources = isset($sources['assigned_sources']) && is_array($sources['assigned_sources'])
+            ? $sources['assigned_sources']
+            : [];
+
+        $this->sourceItemsConfigurationProcessor->process($product->getSku(), $assignedSources);
+    }
+}
