+++ app/code/Magento/InventoryLowQuantityNotificationAdminUi/README.md
@@ -1,0 +1,1 @@
+# InventoryLowQuantityNotificationAdminUi
