+++ app/code/Magento/InventoryLowQuantityNotificationAdminUi/composer.json
@@ -1,18 +1,18 @@
 {
-    "name": "magento/module-inventory-low-quantity-notification",
+    "name": "magento/module-inventory-low-quantity-notification-admin-ui",
     "description": "N/A",
     "require": {
         "php": "~7.1.3||~7.2.0",
         "magento/framework": "*",
         "magento/module-catalog": "*",
-        "magento/module-eav": "*",
+        "magento/module-backend": "*",
+        "magento/module-reports": "*",
         "magento/module-catalog-inventory": "*",
-        "magento/module-inventory": "*",
         "magento/module-inventory-api": "*",
-        "magento/module-inventory-catalog-api": "*",
         "magento/module-inventory-configuration-api": "*",
         "magento/module-inventory-low-quantity-notification-api": "*",
-        "magento/module-store": "*"
+        "magento/module-inventory-low-quantity-notification": "*",
+        "magento/module-ui": "*"
     },
     "type": "magento2-module",
     "license": [
@@ -24,7 +24,7 @@
             "registration.php"
         ],
         "psr-4": {
-            "Magento\\InventoryLowQuantityNotification\\": ""
+            "Magento\\InventoryLowQuantityNotificationAdminUi\\": ""
         }
     }
 }
