+++ app/code/Magento/InventoryLowQuantityNotificationAdminUi/etc/adminhtml/di.xml
@@ -1,0 +1,26 @@
+<?xml version="1.0"?>
+<!--
+/**
+ * Copyright © Magento, Inc. All rights reserved.
+ * See COPYING.txt for license details.
+ */
+-->
+<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:ObjectManager/etc/config.xsd">
+    <virtualType name="Magento\Catalog\Ui\DataProvider\Product\Form\Modifier\Pool">
+        <arguments>
+            <argument name="modifiers" xsi:type="array">
+                <item name="config" xsi:type="array">
+                    <item name="class" xsi:type="string">Magento\InventoryLowQuantityNotificationAdminUi\Ui\DataProvider\Product\Form\Modifier\SourceItemConfiguration</item>
+                    <item name="sortOrder" xsi:type="number">300</item>
+                </item>
+            </argument>
+        </arguments>
+    </virtualType>
+    <type name="Magento\Framework\App\Rss\RssManagerInterface">
+        <arguments>
+            <argument name="dataProviders" xsi:type="array">
+                <item name="notifystock" xsi:type="string">Magento\InventoryLowQuantityNotificationAdminUi\Block\Adminhtml\Rss\NotifyStock</item>
+            </argument>
+        </arguments>
+    </type>
+</config>
