+++ app/code/Magento/InventoryLowQuantityNotificationAdminUi/etc/module.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0"?>
+<!--
+/**
+ * Copyright © Magento, Inc. All rights reserved.
+ * See COPYING.txt for license details.
+ */
+-->
+<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:Module/etc/module.xsd">
+    <module name="Magento_InventoryLowQuantityNotificationAdminUi" setup_version="1.0.0" >
+        <sequence>
+            <module name="Magento_Catalog"/>
+            <module name="Magento_Inventory"/>
+            <module name="Magento_InventoryCatalog"/>
+        </sequence>
+    </module>
+</config>
