+++ app/code/Magento/InventoryLowQuantityNotificationAdminUi/registration.php
@@ -1,0 +1,12 @@
+<?php
+/**
+ * Copyright © Magento, Inc. All rights reserved.
+ * See COPYING.txt for license details.
+ */
+declare(strict_types=1);
+
+\Magento\Framework\Component\ComponentRegistrar::register(
+    \Magento\Framework\Component\ComponentRegistrar::MODULE,
+    'Magento_InventoryLowQuantityNotificationAdminUi',
+    __DIR__
+);
