+++ app/code/Magento/InventoryLowQuantityNotificationAdminUi/view/adminhtml/layout/reports_report_product_lowstock.xml
