+++ app/code/Magento/InventoryLowQuantityNotificationAdminUi/view/adminhtml/web/js/components/use-config-settings.js
